pyproject.toml
[tool.poetry]
name = "frames2osb"
version = "0.1.0"
description = "Turns (video) frames to osu!storyboard"
authors = ["Rendy Arya Kemal <rendyarya22@gmail.com>"]
license = "MIT"
build = 'build.py'

[tool.poetry.dependencies]
python = ">3.8,<3.11"
Pillow = "^8.3.1"
numpy = "^1.21.2"
typed-argument-parser = "^1.7.1"

[tool.poetry.dev-dependencies]
mypy = "^0.910"
flake8 = "^3.9.2"
black = {version = "^21.7b0", allow-prereleases = true}
types-Pillow = "^8.3.3"

[build-system]
requires = ["poetry-core>=1.0.0", "mypy"]
build-backend = "poetry.core.masonry.api"
